Create manager and transfert delete method from repo to manager
