+++ app/public/index.php
@@ -7,6 +7,7 @@ use App\Controller\Admin\AdminPostController;
 use App\Controller\Front\HomeController;
 use App\Controller\Front\PostController;
 use App\Model\Repository\PostRepository;
+use App\Model\Manager\PostManager;
 
 $loader = new \Twig\Loader\FilesystemLoader('../templates/');
 $twig = new \Twig\Environment($loader, [
@@ -20,6 +21,8 @@ try {
     $pdo = new \PDO("mysql:host=" . $config['db_host'] . ";dbname=" . $config['db_name'] . ";charset=utf8", $config['db_user'], $config['db_pwd']);
     $pdo->setAttribute(\PDO::ATTR_ERRMODE, \PDO::ERRMODE_EXCEPTION);
     $postRepository = new PostRepository($pdo);
+    $postManager = new PostManager($pdo);
+
 
     // Front
     if (isset($_GET['post'])) {
@@ -32,7 +35,7 @@ try {
 
         // Back
     } elseif (isset($_GET['admin-post'])) {
-        $adminPostController = new AdminPostController($twig, $postRepository);
+        $adminPostController = new AdminPostController($twig, $postRepository, $postManager);
         if ($_GET['admin-post'] <= 0) {
             $adminPostController->index();
         } else {
