+++ app/src/controller/Admin/AdminPostController.php
@@ -21,7 +21,7 @@ class AdminPostController extends Controller
     public function delete($id_post)
     {
         if ($this->repository->getUniqueById((int)$id_post)) {
-            if ($this->repository->delete((int)$id_post)) {
+            if ($this->manager->delete((int)$id_post)) {
                 $message = 'Le post n° ' . $id_post . ' a été supprimé';
             } else {
                 $message = 'Impossible de supprimer le post n° ' . $id_post . ' !';
