+++ app/src/controller/Controller.php
@@ -8,10 +8,12 @@ abstract class Controller
 {
     protected Environment $twig;
     protected $repository;
+    protected $manager;
 
-    public function __construct(Environment $twig, $repository = null)
+    public function __construct(Environment $twig, $repository = null, $manager = null)
     {
         $this->twig = $twig;
         $this->repository = $repository;
+        $this->manager = $manager;
     }
 }
